Fix app/page.tsx and make app/view/page.tsx for api tests
