+++ client/package.json
@@ -16,9 +16,9 @@
     "@tanstack/react-virtual": "^3.13.4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "lucide-react": "^0.503.0",
     "echart": "^0.1.3",
     "echarts-for-react": "^3.0.2",
+    "lucide-react": "^0.503.0",
     "next": "^15.3.1",
     "papaparse": "^5.5.2",
     "react": "^19.0.0",
